ETL/requirements.txt
pandas
apache-airflow-providers-postgres
psycopg2-binary
gc-python-utils